pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "voice recognition"
authors = ["kerobot"]

[tool.poetry.dependencies]
python = "^3.6"
PyAudio = "^0.2.11"
pylint = "^2.6.0"
numpy = "^1.19.2"
requests = "^2.24.0"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
